Q2/requirements.txt
asgiref==3.7.1
Django==4.2.1
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
drf-yasg==1.21.5